+++ package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "@getbrevo/brevo": "^2.2.0",
     "@vercel/kv": "^2.0.0",
+    "@ybot1122/toby-ui": "^0.0.2-alpha",
     "framer-motion": "^11.2.9",
     "next": "14.2.3",
     "react": "^18",
